Para aplicar control de usuario/fecha en flag

+++ controller/manage/projects.php
@@ -56,16 +56,14 @@ namespace Goteo\Controller\Manage {
                 Model\Contract::setStatus($id, array($subaction => 0));
                 
                 // si están quitando pdf, eliminamos el registro de contrato y archivo
-                /*
-                $contract = Model\Contract::get($id);
-                Model\Contract::setPdf($id, '');
-                $pdf_name = 'contrato-goteo_'.$contract->fullnum . '.pdf';
-                $filename = Model\Contract\Document::$dir . $id . '/' . $pdf_name;
-                if (file_exists($filename)) {
-                    unlink($filename);
+                if ($subaction == 'pdf') {
+                    list($num, $cdate) = Model\Contract::getNum($id);
+                    $pdf_name = 'contrato-goteo_'.$num.'-'.$cdate.'.pdf';
+                    $filename = Model\Contract\Document::$dir . $id . '/' . $pdf_name;
+                    if (file_exists($filename)) {
+                        unlink($filename);
+                    }
                 }
-                 */
-                
                 
                 throw new Redirection('/manage/projects/#'.$id);
             }
